Adding all of Yuan's Code into the more general "code" folder
This is the first step in all using the same enviornment for the
analysis. I have moved all of Yuan's code into the dropbox folder I
shared so we can use GitHub. The code itself hasn't been updated yet to
run from these folders and we may want to create sub-directories in the
future to organize so many .do files.